restoreobjects: use new wrapper for test
